fixed up a bunch of work pages and unpublished a couple

+++ glsl/basic.vert
@@ -2,7 +2,7 @@ precision highp float;
 
 attribute vec3 aPosition;
 uniform vec2 uResolution;
-uniform float uPixelDensity;
+uniform lowp float uPixelDensity;
 
 void main() {
     gl_Position = vec4(aPosition*uPixelDensity,1.0);
